rename(IT-Cotato#185): 패키지 구조 변경
