+++ src/main/java/com/cotato/kampus/domain/board/implement/BoardAppender.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.cotato.kampus.domain.board.dao.entity.BoardEntity;
-import com.cotato.kampus.domain.board.dao.repository.BoardRepository;
+import com.cotato.kampus.domain.board.implement.port.BoardRepository;
 import com.cotato.kampus.domain.board.enums.BoardStatus;
 import com.cotato.kampus.domain.board.enums.BoardType;
