+++ src/main/java/com/cotato/kampus/domain/board/implement/BoardFavoriteReader.java
@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.cotato.kampus.domain.board.dao.repository.BoardFavoriteRepository;
+import com.cotato.kampus.domain.board.implement.port.BoardFavoriteRepository;
 import com.cotato.kampus.domain.board.dao.entity.BoardFavoriteEntity;
 
 import lombok.AccessLevel;
